mysite/requirements.txt
django >=1.11.0
celery>=4.0
celery[redis]
mysqlclient
django-reversion